nag_opt_nlp!: by default, transpose the contraint matrix argument.
I was under the misapprehension that NAG expects this matrix in
column-major order, but that is actually not the case: NAG uses
row-major order here. The `transpose` keyword argument lets you
control whether A gets transposed or not. When `transpose=true`
the rows of A are linear constraints; when `transpose=false` it
is the columsn that are linear constraints.