nag_opt_lp!: don't pass a comm argument (bugfix, this was wrong).
This was causing memory corruption -- the comm argument to the NAG
nag_opt_lp function is not the same kind of object as the one by
the same name to the nag_1d_quad_inf_1 function. This one has many
fields, which were being written to during the execution of the LP
optimization, overwriting the single-word buffer allocate for it.
Since we aren't using this facility, the simple solution is simply
to pass NULL for the comm argument instead.